Move custom endpoint classes to own package
