+++ plugins/src/main/java/org/smartregister/fhir/gateway/plugins/endpoint/BaseEndpoint.java
@@ -1,14 +1,17 @@
-package org.smartregister.fhir.gateway.plugins;
+package org.smartregister.fhir.gateway.plugins.endpoint;
+
+import java.io.IOException;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.parser.IParser;
+import org.smartregister.fhir.gateway.plugins.RestUtils;
+
 import com.google.fhir.gateway.TokenVerifier;
 
-import java.io.IOException;
+import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.parser.IParser;
 
 public abstract class BaseEndpoint extends HttpServlet {
     protected final TokenVerifier tokenVerifier = TokenVerifier.createFromEnvVars();
@@ -19,6 +22,6 @@ public abstract class BaseEndpoint extends HttpServlet {
     protected void doOptions(HttpServletRequest request, HttpServletResponse response) {
         RestUtils.addCorsHeaders(response);
     }
-    protected BaseEndpoint() throws IOException {
-    }
+
+    protected BaseEndpoint() throws IOException {}
 }
