+++ plugins/src/main/java/org/smartregister/fhir/gateway/plugins/endpoint/PractitionerDetailEndpoint.java
@@ -1,4 +1,4 @@
-package org.smartregister.fhir.gateway.plugins;
+package org.smartregister.fhir.gateway.plugins.endpoint;
 
 import static org.smartregister.fhir.gateway.plugins.Constants.KEYCLOAK_UUID;
 
@@ -10,6 +10,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.http.HttpStatus;
+import org.smartregister.fhir.gateway.plugins.Constants;
+import org.smartregister.fhir.gateway.plugins.PractitionerDetailsEndpointHelper;
+import org.smartregister.fhir.gateway.plugins.RestUtils;
+import org.smartregister.fhir.gateway.plugins.Utils;
 import org.smartregister.model.practitioner.PractitionerDetails;
 
 import com.google.fhir.gateway.TokenVerifier;
